+++ MovieTopia/MaintainSeats.cs
@@ -28,9 +28,10 @@ namespace MovieTopia
         }
         private void Form_Resize(Object sender, EventArgs e)
         {
-            
-            // Keep the other controls as they are
-            lblTheater_num.Left = (this.ClientSize.Width - lblTheater_num.Width) / 2;
+
+            // set buttons and labels accordingly for form resizing
+            lblName.Top = padding / 2;
+            lblName.Left = (this.ClientSize.Width - lblName.Width) / 2;
             btnEdit.Left = (this.ClientSize.Width - btnEdit.Width) / 2;
             btnNew.Left = btnEdit.Left - btnEdit.Width - padding;
             btnDelete.Left = btnEdit.Left + btnEdit.Width + padding;
@@ -43,13 +44,10 @@ namespace MovieTopia
             btnReturn.Top = (this.ClientSize.Height - btnDelete.Height - padding * 2);
 
             cbTeater_Name.Left = (this.ClientSize.Width - cbTeater_Name.Width - padding * 2);
-            cbTeater_Name.Top = lblTheater_num.Top + lblTheater_num.Height;
+            cbTeater_Name.Top = lblName.Top + lblName.Height;
             AdjustDataGridViewSize();
             AdjustColumnWidths();
         }
-
-
-
 
         private void LoadData()
         {
@@ -217,6 +215,9 @@ namespace MovieTopia
             {
                 DataGridViewRow selectedRow = dgvSeat.SelectedRows[0];
 
+                DialogResult confirm = MessageBox.Show($"Are you sure you want to delete \"{selectedRow.Cells["SeatColumn"].Value}{selectedRow.Cells["SeatRow"].Value}\"?", "Delete Seat", MessageBoxButtons.YesNo);
+                if (confirm == DialogResult.No) return;
+
                 string sql = @"
                         DELETE FROM 
                             Seat
@@ -238,6 +239,10 @@ namespace MovieTopia
                         connection.Open();
                         command.ExecuteNonQuery();
                         MessageBox.Show("Deleted Successfully", "Success");
+                    }
+                    catch (SqlException)
+                    {
+                        MessageBox.Show($"\"{selectedRow.Cells["SeatColumn"].Value}{selectedRow.Cells["SeatRow"].Value}\" cannot be deleted because it is being referenced by 1 or more Tickets for movies on display, or up-and-coming movies.", "Error");
                     }
                     catch (Exception ex)
                     {
