+++ MovieTopia/MaintainTheatres.cs
@@ -31,6 +31,7 @@ namespace MovieTopia
 
         private void Form_Resize(Object sender, EventArgs e)
         {
+            lblName.Top = padding / 2;
             lblName.Left = (this.ClientSize.Width - lblName.Width) / 2;
             btnEdit.Left = (this.ClientSize.Width - btnEdit.Width) / 2;
             btnNew.Left = btnEdit.Left - btnEdit.Width - padding;
@@ -207,6 +208,9 @@ namespace MovieTopia
             {
                 DataGridViewRow selectedRow = dgvData.SelectedRows[0];
 
+                DialogResult confirm = MessageBox.Show($"Are you sure you want to delete \"{selectedRow.Cells["TheatreName"].Value}\"?", "Delete Theatre", MessageBoxButtons.YesNo);
+                if (confirm == DialogResult.No) return;
+
                 string sql = @"
                         DELETE FROM 
                             Theatre
@@ -228,6 +232,10 @@ namespace MovieTopia
                         connection.Open();
                         command.ExecuteNonQuery();
                         MessageBox.Show("Deleted Successfully", "Success");
+                    }
+                    catch (SqlException)
+                    {
+                        MessageBox.Show($"\"{selectedRow.Cells["TheatreName"].Value}\" cannot be deleted because of 1 or more Scheduled Movies that are dependent on this Theatre. Please delete the corresponding records before attempting to delete this Theatre.", "Error");
                     }
                     catch (Exception ex)
                     {
